refactor: prepare data_dir structure for a future update
prepare data_dir structure for a future update when poof will be able to
support multiple services apart github.com. This will avoid having users
to migrate once again in the future, and will allow us to understand
where an app has be installed from, as the same app could be available
in and fetched from multiple repositories.